+++ Bugsnag/Assets/Bugsnag/Editor/SymbolUpload/BugsnagSymbolUploader.cs
@@ -34,7 +34,7 @@ fi
                 return;
             }
 
-            var config = BugsnagSettingsObject.GetSettingsObject();
+            var config = BugsnagSettingsObject.LoadBuildTimeSettingsObject();
             if (config == null || !config.AutoUploadSymbols)
             {
                 return;
@@ -134,10 +134,10 @@ fi
         string GetMacosXcodeProjectPath(string outputPath)
         {
             string[] parts = outputPath.Split('/');
-            string xcprojFile = parts[^1] + ".xcodeproj";
+            string xcprojFile = parts[parts.Length - 1] + ".xcodeproj";
             return outputPath + "/" + xcprojFile + "/project.pbxproj";
         }
-        
+
         private string RemoveShellScriptPhase(string project, string guid)
         {
             // Search for and remove the phase object from the XML. only match the guid followed by the braces
@@ -155,7 +155,7 @@ fi
         private string GetDsymUploadCommand(BugsnagSettingsObject config)
         {
             var cli = new BugsnagCLI();
-            var command = cli.GetIosDsymUploadCommand(config.ApiKey, config.UploadEndpoint, config.AppVersion);
+            var command = cli.GetIosDsymUploadCommand(config.ApiKey, config.UploadEndpoint);
             return DSYM_UPLOAD_SCRIPT_TEMPLATE.Replace("<CLI_COMMAND>", command);
         }
     }