+++ Bugsnag/Assets/Bugsnag/Runtime/BugsnagSettingsObject.cs
@@ -58,9 +58,14 @@ namespace BugsnagUnity
         public int SwitchCacheIndex = 0;
         public int SwitchCacheMaxSize = 10485760;
 
-        internal static BugsnagSettingsObject GetSettingsObject()
+        public static BugsnagSettingsObject LoadBuildTimeSettingsObject()
         {
-            return Resources.Load<BugsnagSettingsObject>("Bugsnag/BugsnagSettingsObject");
+            var settings = Resources.Load<BugsnagSettingsObject>("Bugsnag/BugsnagSettingsObject");
+            if (settings == null)
+            {
+                throw new Exception("No BugsnagSettingsObject found.");
+            }
+            return settings;
         }
 
         public static Configuration LoadConfiguration()