+++ Bugsnag/Assets/Bugsnag/Runtime/Payload/Session.cs
@@ -74,13 +74,13 @@ namespace BugsnagUnity.Payload
 
         internal void AddException(Report report)
         {
-            if (report.IsHandled)
+            if(report.IsHandled)
             {
-                Events.IncrementHandledCount();
+                Events.UpdateHandledCount(true);
             }
             else
             {
-                Events.IncrementUnhandledCount();
+                Events.UpdateUnhandledCount(true);
             }
         }
 
@@ -96,32 +96,36 @@ namespace BugsnagUnity.Payload
 
     internal class SessionEvents : Dictionary<string, int>
     {
-        private readonly object _handledLock = new object();
-        private readonly object _unhandledLock = new object();
+        private readonly object _lock = new object();
+
+        private const string HANDLED_KEY = "handled";
+        private const string UNHANDLED_KEY = "unhandled";
 
         internal SessionEvents(int handled, int unhandled)
         {
-            this.AddToPayload("handled", handled);
-            this.AddToPayload("unhandled", unhandled);
+            this.AddToPayload(HANDLED_KEY, handled);
+            this.AddToPayload(UNHANDLED_KEY, unhandled);
         }
 
-        internal int Handled => this["handled"];
-        internal int Unhandled => this["unhandled"];
+        internal int Handled => this[HANDLED_KEY];
+        internal int Unhandled => this[UNHANDLED_KEY];
 
-        internal void IncrementHandledCount()
+        private void UpdateEventCount(bool handled, bool increment)
         {
-            lock (_handledLock)
+            lock (_lock)
             {
-                this["handled"]++;
+                this[handled ? HANDLED_KEY : UNHANDLED_KEY] += increment ? 1 : -1;
             }
         }
 
-        internal void IncrementUnhandledCount()
+        internal void UpdateUnhandledCount(bool increment)
         {
-            lock (_unhandledLock)
-            {
-                this["unhandled"]++;
-            }
+            UpdateEventCount(false, increment);
+        }
+
+        internal void UpdateHandledCount(bool increment)
+        {
+            UpdateEventCount(true, increment);
         }
     }
 }