+++ features/scripts/build_android.sh
@@ -23,7 +23,7 @@ popd
 pushd "$script_path/../fixtures"
 
 # Run unity and immediately exit afterwards, log all output
-DEFAULT_CLI_ARGS="-quit -batchmode -nographics -logFile build_android_apk.log"
+DEFAULT_CLI_ARGS="-quit -batchmode -nographics -logFile build_android_apk.log -buildTarget Android"
 
 project_path=`pwd`/maze_runner